graphics: base: Import lib[g]scaler from Android 10 BSP sources
as legacy_libscaler and legacy_libgscaler and put them behind a flag,
BOARD_USES_LEGACY_LIBSCALER.

The new libscaler heavily depends on libacryl. On some older devices,
which will be relevant when HWC1 support will be brought back, libacryl
isn't used, and when it is present, it's just there as a dummy blob
without a scaler/compositor/blter set as per decompilations, and
attempts at setting them and trying to use libacryl in any real way
don't work. The old libscaler remains relevant to them and properly
working.

Also, on 7885/7904 devices with the new libscaler, after ending a screen
recording:
 libacryl: Failed VIDIOC_S_FMT .type=10, .width=1008, .height=2160, .pixelformat=0x34424752 [Inappropriate ioctl for device]
This error doesn't occur with the old libscaler.
Note also that the libacryl config (scaler/compositor) matches stock
blobs as per decompilations, and it otherwise works.

As for libgscaler, Android 10 and older camera blobs depend on it, and
later on, when again also bringing back HWC1 support, it will also
depend on it.

Change-Id: Id0f159504c08821328e6cdf1daae61b2e590cbc1
Signed-off-by: HeroBuxx <me@herobuxx.me>